rename package and classes: calendar -> monthly
