+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
       android:label="@string/app_name"
       android:supportsRtl="true"
       android:theme="@style/AppTheme">
-    <activity android:name=".calendar.CalendarViewActivity"
+    <activity android:name=".monthly.MonthlyViewActivity"
         android:launchMode="singleTask">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
