+++ app/src/main/java/io/github/yusukeiwaki/u1/monthly/CalendarDayListHelper.java
@@ -1,4 +1,4 @@
-package io.github.yusukeiwaki.u1.calendar;
+package io.github.yusukeiwaki.u1.monthly;
 
 import com.annimon.stream.Stream;
 import java.util.List;
