+++ app/src/main/java/io/github/yusukeiwaki/u1/monthly/CalendarDayViewHolder.java
@@ -1,4 +1,4 @@
-package io.github.yusukeiwaki.u1.calendar;
+package io.github.yusukeiwaki.u1.monthly;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
